Fix build on case-sensitive filesystems
Add built artifacts (bin/) to gitignore.

Un-hardcode g++, defaulting to 'c++' if not overriden by the environment.

Go ahead and enable some warings, but not -Werror.

Simplify/cleanup makefile, miscellaneous.

Optimize for local CPU for now.  Not suitable for producing distributed
binaries but I don't know that anyone does that.